# examples/v2/open_api/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Sample Swagger App
# During dev, should point to your local machine
# Deployment Manager will need a public endpoint to make requests, if you dont set host
# the URL where this swagger.yaml is served is going to be used as host.
# If you specify port here, you can only use http or https.
# If you dont specify port, 80 is going to be used for http, and 443 is going to be used for https.
# host: "[api-name].endpoints.[project-id].cloud.goog"
host: api-sample.endpoints.sondrago-blue.cloud.goog
# basePath prefixes all resource paths
basePath: /
#
schemes:
  # If there is http and https Deployment manager will prefer https
  - https
# format of bodies a client can send (Content-Type)
consumes:
# DM only support application/json APIs
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  # Name of the REST resource is defined in the body.
  /body_named_resource:
    # binds a127 app logic to a route
    x-swagger-router-controller: named_resource
    post:
      description: Creates a resource, name comes in the body
      # used as the method name of the controller
      operationId: createBodyNamedResource
      parameters:
        - $ref: "#/parameters/namedResource"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/NamedResource"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /body_named_resource/{resourceName}:
    x-swagger-router-controller: named_resource
    get:
      description: Returns 'resource' to the caller
      # used as the method name of the controller
      operationId: getBodyNamedResource
      parameters:
        - $ref: "#/parameters/resourceName"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/NamedResource"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: Updates a body named resource
      # used as the method name of the controller
      operationId: updateBodyNamedResource
      parameters:
        - $ref: "#/parameters/namedResource"
        - $ref: "#/parameters/resourceName"

      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/NamedResource"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: Deletes a body named resource
      # used as the method name of the controller
      operationId: deleteBodyNamedResource
      parameters:
        - $ref: "#/parameters/resourceName"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Empty"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /server_named_resource:
    # binds a127 app logic to a route
    x-swagger-router-controller: named_resource
    post:
      description: Creates a resource, server sets the name
      # used as the method name of the controller
      operationId: createServerNamedResource
      parameters:
        - $ref: "#/parameters/namedResource"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/NamedResource"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /server_named_resource/{resourceName}:
    x-swagger-router-controller: named_resource
    get:
      description: Returns 'resource' to the caller
      # used as the method name of the controller
      operationId: getServerNamedResource
      parameters:
        - $ref: "#/parameters/resourceName"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/NamedResource"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: Updates a server named resource
      # used as the method name of the controller
      operationId: updateServerNamedResource
      parameters:
        - $ref: "#/parameters/namedResource"
        - $ref: "#/parameters/resourceName"

      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/NamedResource"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: Deletes a named resource
      # used as the method name of the controller
      operationId: deleteServerNamedResource
      parameters:
        - $ref: "#/parameters/resourceName"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Empty"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /named_resource/{resourceName}:
    x-swagger-router-controller: named_resource
    get:
      description: Returns 'Resource' to the caller
      # used as the method name of the controller
      operationId: getNamedResource
      parameters:
        - $ref: "#/parameters/resourceName"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/NamedResource"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Creates a named resource
      # used as the method name of the controller
      operationId: createNamedResource
      parameters:
        - $ref: "#/parameters/namedResource"
        - $ref: "#/parameters/resourceName"

      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/NamedResource"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: Updates a named resource
      # used as the method name of the controller
      operationId: updateNamedResource
      parameters:
        - $ref: "#/parameters/namedResource"
        - $ref: "#/parameters/resourceName"

      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/NamedResource"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: Deletes a named resource
      # used as the method name of the controller
      operationId: deleteNamedResource
      parameters:
        - $ref: "#/parameters/resourceName"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Empty"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
    get:
      description: Returns 'Resource' to the caller
      # used as the method name of the controller
      operationId: getSwagger
      responses:
        "200":
          description: Success
          schema:
            type: object
            properties: {}
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
# Complex objects have schema definitions
definitions:
  NamedResource:
    required:
      - message
    properties:
      name:
        type: string
      message:
        type: string
      fingerprint:
        type: string
  Empty:
    properties: {}
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
# Common parameters
parameters:
  namedResource:
    name: body
    in: body
    description: The named resource
    required: true
    schema:
      $ref: "#/definitions/NamedResource"
  resourceName:
    name: resourceName
    in: path
    description: The name of the resource created
    required: true
    type: string